ich9: add TCO interface emulation
This interface is provides some registers within a 32-byte range and can
be accessed through PCI-to-LPC bridge interface (PMBASE + 0x60).

It's commonly used as a watchdog timer to detect system lockups through
SMIs that are generated by it (if TCO_EN bit is enabled) on every
timeout and then giving the system firmware the ability to perform any
action -- if NO_REBOOT bit is not set in GCS (General Control and Status
register), the system will be reseted upon second TCO timeout.

This patch focus on adding full support to TCO watchdog logic and a few
features like mapping NMIs to SMIs (NMI2SMI_EN bit), system intruder
detection (TCO interrupt due to INTRUDER# signal), etc. are not
implemented yet.

Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>